Fix windows build warnings
Most of these were the VS compiler being overly picky about types (or
just rubbish at static analysis), but it did make me notice a subtle and
nasty little bug in handling datetimes with negative TZ offsets so
worthwhile overall!

+++ source/decoder.h
@@ -11,7 +11,7 @@ typedef struct {
     PyObject *shareables;
     PyObject *str_errors;
     bool immutable;
-    int32_t shared_index;
+    Py_ssize_t shared_index;
 } CBORDecoderObject;
 
 PyTypeObject CBORDecoderType;
